refactor: 로그아웃 관련 코드 분리
